Refactor: PR 리뷰 반연
Ref: #36
