+++ src/test/java/com/studypals/domain/studyManage/service/StudyCategoryServiceTest.java
@@ -22,7 +22,6 @@ import com.studypals.domain.studyManage.dto.UpdateCategoryReq;
 import com.studypals.domain.studyManage.dto.mappers.CategoryMapper;
 import com.studypals.domain.studyManage.entity.StudyCategory;
 import com.studypals.domain.studyManage.worker.StudyCategoryReader;
-import com.studypals.domain.studyManage.worker.StudyCategoryWriter;
 import com.studypals.global.exceptions.errorCode.StudyErrorCode;
 import com.studypals.global.exceptions.exception.StudyException;
 
@@ -37,9 +36,6 @@ class StudyCategoryServiceTest {
 
     @Mock
     private MemberReader memberReader;
-
-    @Mock
-    private StudyCategoryWriter studyCategoryWriter;
 
     @Mock
     private StudyCategoryReader studyCategoryReader;
@@ -170,7 +166,7 @@ class StudyCategoryServiceTest {
         Long categoryId = 1L;
         UpdateCategoryReq req = new UpdateCategoryReq(categoryId, "new category", "#FFFFFF", 12, "new description");
 
-        given(studyCategoryWriter.findAndValidate(userId, categoryId)).willReturn(mockStudyCategory);
+        given(studyCategoryReader.findAndValidate(userId, categoryId)).willReturn(mockStudyCategory);
         given(mockStudyCategory.getId()).willReturn(categoryId);
 
         // when
